+++ app/src/main/cpp/skyline/services/nvdrv/INvDrvServices.h
@@ -5,63 +5,18 @@
 
 #include <services/base_service.h>
 #include <services/serviceman.h>
-#include <kernel/types/KTransferMemory.h>
-#include <gpu.h>
-#include "devices/nvdevice.h"
-#include "devices/nvhost_syncpoint.h"
 
 namespace skyline::service::nvdrv {
+    class Driver;
+
     /**
      * @brief nvdrv or INvDrvServices is used to access the Nvidia GPU inside the Switch (https://switchbrew.org/wiki/NV_services#nvdrv.2C_nvdrv:a.2C_nvdrv:s.2C_nvdrv:t)
      */
     class INvDrvServices : public BaseService {
       private:
-        std::unordered_map<device::NvDeviceType, std::shared_ptr<device::NvDevice>> deviceMap; //!< A map from a NvDeviceType to the NvDevice object
-        std::unordered_map<u32, std::shared_ptr<device::NvDevice>> fdMap; //!< A map from an FD to a shared pointer to it's NvDevice object
-        u32 fdIndex{}; //!< Holds the index of a file descriptor
-
-        /**
-         * @brief Open a specific device and return a FD
-         * @param path The path of the device to open an FD to
-         * @return The file descriptor to the device
-         */
-        u32 OpenDevice(const std::string &path);
-
-        /**
-         * @brief Returns a particular device with a specific FD
-         * @tparam objectClass The class of the device to return
-         * @param fd The file descriptor to retrieve
-         * @return A shared pointer to the device
-         */
-        template<typename objectClass>
-        std::shared_ptr<objectClass> GetDevice(u32 fd) {
-            try {
-                auto item = fdMap.at(fd);
-                return std::static_pointer_cast<objectClass>(item);
-            } catch (std::out_of_range) {
-                throw exception("GetDevice was called with invalid file descriptor: 0x{:X}", fd);
-            }
-        }
+        std::shared_ptr<Driver> driver;
 
       public:
-        NvHostSyncpoint hostSyncpoint;
-
-        /**
-         * @brief Returns a particular device with a specific type
-         * @tparam objectClass The class of the device to return
-         * @param type The type of the device to return
-         * @return A shared pointer to the device
-         */
-        template<typename objectClass>
-        std::shared_ptr<objectClass> GetDevice(device::NvDeviceType type) {
-            try {
-                auto item = deviceMap.at(type);
-                return std::static_pointer_cast<objectClass>(item);
-            } catch (std::out_of_range) {
-                throw exception("GetDevice was called with invalid type: 0x{:X}", type);
-            }
-        }
-
         INvDrvServices(const DeviceState &state, ServiceManager &manager);
 
         /**
