+++ app/src/main/cpp/skyline/services/nvdrv/devices/nvdevice.h
@@ -10,6 +10,8 @@
 #define NFUNC(function) std::bind(&function, this, std::placeholders::_1)
 
 namespace skyline::service::nvdrv::device {
+    using namespace kernel;
+
     /**
      * @brief An enumeration of all the devices that can be opened by nvdrv
      */
